+++ components/pages/home/HomePage.tsx
@@ -6,6 +6,7 @@ import { GiHamburgerMenu, GiCancel } from 'react-icons/gi';
 import OnGoingTimerArea from '../../../features/workSession/components/elements/OnGoingTimerArea/OnGoingTimerArea';
 import TabsArea from '../../../features/workSession/components/elements/TabArea/TabsArea';
 import LoadingOverlay from '../../elements/common/LoadingOverlay/LoadingOverlay';
+import MobileMenu from '../../elements/common/MobileMenu/MobileMenu';
 
 import Navbar from '../../elements/Navbar/Navbar';
 
@@ -94,88 +95,14 @@ const MainAreaContainer = styled.div`
   width: 100%;
   height: 100%;
   gap: 24px;
-  padding-inline: 24px;
-  padding-bottom: 24px;
+  padding: 2em 1.5em 1.5em;
 
   @media ${breakPoint.tablet} {
     flex-direction: row;
+    padding-inline: 24px;
+    padding-bottom: 24px;
   }
 `;
-
-
-
-// Styled components for the menu and overlay
-const MenuContainer = styled.div<{isOpen: boolean}>`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
-  display: flex;
-  /* align-items: center; */
-  justify-content: center;
-  opacity: ${(props) => (props.isOpen ? '1' : '0')};
-  pointer-events: ${(props) => (props.isOpen ? 'auto' : 'none')};
-  transition: opacity 0.3s ease;
-  z-index: 1000;
-`;
-
-const MenuContent = styled.div`
-  height: 40%;
-  background-color: #fff;
-  padding: 20px;
-  border-radius: 8px;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.3);
-  position: absolute;
-  top: 10px; /* Adjust the top position as needed */
-  /* left: 10px; */
-`;
-
-const HamburgerButton = styled.button`
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  background: none;
-  border: none;
-  font-size: 24px;
-  color: #fff;
-  cursor: pointer;
-  z-index: 1100;
-`;
-interface HamburgerMenuProps {
-  items: React.ReactNode[];
-}
-
-const HamburgerMenu: React.FC<HamburgerMenuProps> = ({ items }) => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
-
-  const closeMenu = () => {
-    setIsOpen(false);
-  };
-
-  return (
-    <>
-      <HamburgerButton onClick={toggleMenu}>
-        {isOpen ? <GiCancel /> : <GiHamburgerMenu />}
-      </HamburgerButton>
-      <MenuContainer isOpen={isOpen} onClick={closeMenu}>
-        <MenuContent onClick={(e) => e.stopPropagation()}>
-          {items.map((item, index) => (
-            <div key={index}>{item}</div>
-          ))}
-        </MenuContent>
-      </MenuContainer>
-    </>
-  );
-};
-
-
-
 
 // You can now use the 'tabs' variable, which is of type 'Tab[]'.
 
@@ -201,7 +128,7 @@ const HomePage = () => {
     <>
       <LoadingOverlay loading={isLoading} />
       <Navbar />
-      <HamburgerMenu items={menuItems} />
+      <MobileMenu items={menuItems} />
       <MainAreaContainer>
         <OnGoingTimerArea />
         <TabsArea tabs={testTabs} />
